+++ client/helpers/findAvgfindSum.js
