+++ client/helpers/isContinuous.js
