+++ client/helpers/isSDLow.js
