+++ client/helpers/tests/sd.spec.js
@@ -1,4 +1,4 @@
-const {findMean, findSD, isSDLow} = require('../sd')
+const {findMean, findSD, isSDLow} = require('../isSDLow')
 
 const data = [1, 4, 6, -100, 9, 2, 0.2]
